requirements.txt
pandas
duckdb
pyarrow
requests
tqdm
streamlit
pydeck
pyyaml
